+++ lib/di.dart
@@ -3,6 +3,8 @@ library di;
 import 'src/provider.dart';
 import 'key.dart';
 
+export 'key.dart' show Key;
+
 part 'src/injector.dart';
 part 'src/module.dart';
 part 'src/errors.dart';