+++ lib/dynamic_injector.dart
@@ -1,10 +1,10 @@
 library di.dynamic_injector;
 
 import 'di.dart';
-import 'key.dart';
 import 'src/mirrors.dart';
 import 'src/base_injector.dart';
 import 'src/error_helper.dart';
+import 'src/provider.dart';
 
 export 'di.dart';