+++ lib/static_injector.dart
@@ -1,9 +1,9 @@
 library di.static_injector;
 
 import 'di.dart';
-import 'key.dart';
 import 'src/error_helper.dart';
 import 'src/base_injector.dart';
+import 'src/provider.dart';
 
 export 'annotations.dart';
 export 'di.dart';